Serialize/Deserialize layout. #6
The Tasks would render on top of each other in the top left
because positional data was not persisted.

This change uses react-diagram's engine's serialize/deserialize
on the componentDidMount and componentDidUpdate to get and set
layout snapshots to localStorage.

This implementation does not consider the case when a task is
added to the datastore but not reflected in local storage. This should
be reconciled for the next sprint